ingest/airbyte/terraform_azure_function/function/csv_to_parquet/requirements.txt
azure-functions
azure-storage-file-datalake>=12.6.0
pandas>=1.3.5
pyarrow>=7.0.0
fastparquet>=0.8.0